- split and combine confirmation - hook for token state
